==> notebooks/01_data_loading.ipynb <==
import pandas as pd


# ---------- Load CSV files from data/raw ----------
btc  = pd.read_csv("../data/raw/BTC-USD.csv")
eth  = pd.read_csv("../data/raw/ETH-USD.csv")
bnb  = pd.read_csv("../data/raw/BNB-USD.csv")
ada  = pd.read_csv("../data/raw/ADA-USD.csv")
doge = pd.read_csv("../data/raw/DOGE-USD.csv")
dot  = pd.read_csv("../data/raw/DOT-USD.csv")
sol  = pd.read_csv("../data/raw/SOL-USD.csv")
xrp  = pd.read_csv("../data/raw/XRP-USD.csv")
ltc  = pd.read_csv("../data/raw/LTC-USD.csv")
bch  = pd.read_csv("../data/raw/BCH-USD.csv")

# ---------- Add crypto name column ----------
btc['crypto']  = 'BTC'
eth['crypto']  = 'ETH'
bnb['crypto']  = 'BNB'
ada['crypto']  = 'ADA'
doge['crypto'] = 'DOGE'
dot['crypto']  = 'DOT'
sol['crypto']  = 'SOL'
xrp['crypto']  = 'XRP'
ltc['crypto']  = 'LTC'
bch['crypto']  = 'BCH'

# ---------- Combine all data ----------
master_df = pd.concat(
    [btc, eth, bnb, ada, doge, dot, sol, xrp, ltc, bch],
    ignore_index=True
)

# ---------- Date handling ----------
master_df['Date'] = pd.to_datetime(master_df['Date'])
master_df.sort_values(by=['crypto', 'Date'], inplace=True)

master_df.head()

master_df.tail()

master_df['crypto'].value_counts()

# ---------- Save processed master file ----------
master_df.to_csv("../data/processed/crypto_master_data.csv", index=False)
