==> notebooks/02_data_preprocessing.ipynb <==
import pandas as pd

df = pd.read_csv("crypto_master_data.csv")
df.head()

df.info()

df.isnull().sum()

master_df = df[
    ['Date', 'Open', 'High', 'Low', 'Close', 'Volume', 'crypto']
]

master_df.rename(columns={
    'Open': 'open_price',
    'High': 'high_price',
    'Low': 'low_price',
    'Close': 'close_price',
    'Volume': 'volume'
}, inplace=True)


master_df.sort_values(by=['crypto', 'Date'], inplace=True)

master_df.set_index('Date', inplace=True)

master_df.head()

master_df.groupby('crypto').size()

master_df.to_csv("crypto_master_cleaned.csv")
